requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
sqlalchemy>=2.0.22
psycopg2-binary>=2.9.9
pgvector>=0.2.3
alembic>=1.12.0
python-dotenv>=1.0.0
httpx>=0.25.0
pytest>=0.1.2
mypy>=1.6.1
black>=23.10.0
flake8>=6.1.0
asyncpg==0.30.0
langchain-community>=0.0.16
sentence-transformers>=2.2.2
langchain-postgres>=0.0.14